+++ NEWS.md
@@ -6,6 +6,7 @@
 - GW sunlit/shade minimum/maximum daily output
 - FMC calculations with basic water balance output
 - New output in growth simulation (biomass values)
+- New function 'moisture_cohortFMCDay'
 
 # Version 1.1.1
 - Root exudation added to carbon balance